+++ packages/soukai-solid/src/testing/lib/stubs/VCardEmail.ts
@@ -1,0 +1,4 @@
+import Model from './VCardEmail.schema';
+
+export default class VCardEmail extends Model {
+}