+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Recursive detection of removed documents in relations.
+
 ## [v0.8.1](https://github.com/koca2000/soukai/releases/tag/v0.8.1) - 2026-08-28
 
 ### Fixed